+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/base/events/AppEvent.kt
@@ -4,7 +4,7 @@ sealed class AppEvent {
 
   class AppbarEvent(val title: String) : AppEvent()
 
-  class ConnectionEvent(val state: Int) : AppEvent()
+  class ConnectionEvent(val state: Long) : AppEvent()
 
-  class NavDrawerEvent(val itemId: Int) : AppEvent()
+  class NavDrawerEvent(val itemId: Long) : AppEvent()
 }
