+++ app/src/main/java/com/sedsoftware/yaptalker/presentation/features/posting/AddMessagePresenter.kt
@@ -48,7 +48,7 @@ class AddMessagePresenter @Inject constructor(
     super.detachView(view)
   }
 
-  fun insertChosenTag(selectionStart: Int, selectionEnd: Int, @Tag tag: Int) {
+  fun insertChosenTag(selectionStart: Int, selectionEnd: Int, @Tag tag: Long) {
     when {
       tag == MessageTagCodes.TAG_LINK -> onLinkTagClicked()
       tag == MessageTagCodes.TAG_VIDEO -> onVideoLinkTagClicked()
@@ -88,7 +88,7 @@ class AddMessagePresenter @Inject constructor(
     viewState.showImagePickerDialog()
   }
 
-  private fun onTagClickedWithSelection(@Tag tag: Int) {
+  private fun onTagClickedWithSelection(@Tag tag: Long) {
     when (tag) {
       MessageTagCodes.TAG_B -> {
         viewState.insertTags(MessageTags.B_OPEN, MessageTags.B_CLOSE)
@@ -102,7 +102,7 @@ class AddMessagePresenter @Inject constructor(
     }
   }
 
-  private fun onTagClickedWithNoSelection(@Tag tag: Int) {
+  private fun onTagClickedWithNoSelection(@Tag tag: Long) {
     when (tag) {
       MessageTagCodes.TAG_B -> {
         if (isBOpened) {
